Application scaffold
Using Rails 5.0.0.1, application was generated without ActiveRecord (as
we're going to be using Mongo + Mongoid).

Also just to note that the application was generated in API mode, which
is a new featured added in Rails 5. While I'd typically have used
something like Sinatra for an API-only backend, I'm curious to see how
Rails in API mode compares.