Stubbing out the checkout controller
Similarly to Basket, I think it's likely that the checkout controller will only be invoked
when we already have a reference to a Shopper, so I've configured it as a nested
route to make that relationship explicit:

/api/shoppers/1/checkout

Should we decide in future that checkout should operate as an independent top-level
resource (i.e. if we wanted to allow guest checkouts, or just found the design too
restrictive for some reason) it's straightforward to make that change.